26 Алгоритмы сжатия

+++ hw-26-compress/src/main/java/ru/tikskit/hw26compress/Compressor.java
@@ -1,6 +1,6 @@
 package ru.tikskit.hw26compress;
 
 public interface Compressor {
-    int[] compress(int[] data);
-    int[] uncompress(int[] data);
+    byte[] compress(byte[] data);
+    byte[] uncompress(byte[] data);
 }
